fixed pixmap scale error: float needed int

+++ qualcoder/report_compare_coder_file.py
@@ -918,7 +918,7 @@ class DialogDualCodedImage(QtWidgets.QDialog):
         if self.pixmap is None:
             return
         self.scale = (self.ui.horizontalSlider.value() + 1) / 100
-        height = self.scale * self.pixmap.height()
+        height = int(self.scale * self.pixmap.height())
         pixmap = self.pixmap.scaledToHeight(height, QtCore.Qt.TransformationMode.FastTransformation)
         pixmap_item = QtWidgets.QGraphicsPixmapItem(pixmap)
         pixmap_item.setPos(0, 0)
